requirements.txt
asgiref==3.7.2
awsebcli==3.20.10
blessed==1.20.0
boto3==1.34.47
botocore==1.31.85
cement==2.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.3
Django==4.2.10
django-dotenv==1.4.2
django-filter==23.5
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.6
jmespath==1.0.1
pathspec==0.10.1
pillow==10.2.0
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.10.0
semantic-version==2.8.5
six==1.16.0
sqlparse==0.4.4
stripe==8.5.0
termcolor==1.1.0
typing_extensions==4.9.0
urllib3==1.26.18
wcwidth==0.1.9
